Fixes for overrides
