+++ components/shared/components/Widget/components/Widget.tsx
@@ -1,24 +1,11 @@
 import { groq } from "next-sanity";
-import {
-  createContext,
-  useCallback,
-  useContext,
-  useEffect,
-  useMemo,
-  useRef,
-  useState,
-} from "react";
+import { createContext, useCallback, useEffect, useRef, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { useDebouncedCallback } from "use-debounce";
 import { getClient } from "../../../../../lib/sanity.client";
 import { withStaticProps } from "../../../../../util/withStaticProps";
-import { WidgetContext } from "../../../../main/layout/layout";
 import { fetchCauseAreasAction } from "../store/layout/actions";
-import { paymentMethodConfigurations } from "../config/methods";
-import { setRecurring } from "../store/donation/actions";
 import { fetchReferralsAction } from "../store/referrals/actions";
 import { State } from "../store/state";
-import { RecurringDonation } from "../types/Enums";
 import { WidgetProps } from "../types/WidgetProps";
 import { Carousel } from "./Carousel";
 import { DonationPane } from "./panes/DonationPane/DonationPane";
@@ -31,9 +18,18 @@ import {
   TooltipLink,
   TooltipWrapper,
 } from "./shared/ProgressBar/ProgressBar.style";
-import { usePrefilledDistribution, usePrefilledSum, useQueryParamsPrefill } from "./hooks";
+import {
+  useAvailablePaymentMethods,
+  useAvailableRecurringOptions,
+  useDefaultPaymentMethodEffect,
+  usePrefilledDistribution,
+  usePrefilledSum,
+  useQueryParamsPrefill,
+  useWidgetScaleEffect,
+} from "./hooks";
 import { useElementHeight } from "../../../../../hooks/useElementHeight";
 import { PrefilledDistribution } from "../../../../main/layout/WidgetPane/WidgetPane";
+import { RecurringDonation } from "../types/Enums";
 
 export const widgetContentQuery = groq`
 ...,
@@ -89,137 +85,17 @@ export const WidgetTooltipContext = createContext<[{ text: string; link?: string
   () => {},
 ]);
 
-/**
- * Determine available recurring options based on the configured payment methods
- */
-const useAvailableRecurringOptions = (paymentMethods: NonNullable<WidgetProps["methods"]>) => {
-  const recurring = useMemo(
-    () =>
-      paymentMethods.some((method) => {
-        const configuration = paymentMethodConfigurations.find(
-          (config) => config.id === method._id,
-        );
-        return configuration?.recurringOptions.includes(RecurringDonation.RECURRING);
-      }),
-    [paymentMethods],
-  );
-
-  const single = useMemo(
-    () =>
-      paymentMethods.some((method) => {
-        const configuration = paymentMethodConfigurations.find(
-          (config) => config.id === method._id,
-        );
-        return configuration?.recurringOptions.includes(RecurringDonation.NON_RECURRING);
-      }),
-    [paymentMethods],
-  );
-
-  return useMemo(() => ({ recurring, single }), [recurring, single]);
-};
-
-/**
- * Determine available payment methods based on the selected recurring option
- */
-const useAvailablePaymentMethods = (paymentMethods: NonNullable<WidgetProps["methods"]>) => {
-  const recurring = useSelector((state: State) => state.donation.recurring);
-
-  const availablePaymentMethods = useMemo(
-    () =>
-      paymentMethods.filter((method) => {
-        const configuration = paymentMethodConfigurations.find(
-          (config) => config.id === method._id,
-        );
-        return configuration?.recurringOptions.includes(recurring);
-      }),
-    [paymentMethods, recurring],
-  );
-
-  return availablePaymentMethods;
-};
-
-/**
- * This effect is used to set the default payment method to single if recurring is not enabled
- */
-const useDefaultPaymentMethodEffect = (paymentMethods: NonNullable<WidgetProps["methods"]>) => {
-  const dispatch = useDispatch();
-  const recurring = useSelector((state: State) => state.donation.recurring);
-
-  const availableRecurringOptions = useAvailableRecurringOptions(paymentMethods);
-
-  useEffect(() => {
-    if (recurring === RecurringDonation.RECURRING && !availableRecurringOptions.recurring) {
-      dispatch(setRecurring(RecurringDonation.NON_RECURRING));
-    }
-  }, [recurring, availableRecurringOptions.recurring, dispatch]);
-};
-
-/**
- * Scale the widget to fit the screen
- */
-const useWidgetScaleEffect = (widgetRef: React.RefObject<HTMLDivElement>, inline: boolean) => {
-  const [widgetContext, setWidgetContext] = useContext(WidgetContext);
-  const [scalingFactor, setScalingFactor] = useState(1);
-  const [scaledHeight, setScaledHeight] = useState(979);
-  const [lastHeight, setLastHeight] = useState(979);
-  const [lastWidth, setLastWidth] = useState(400);
-
-  const scaleWidget = useCallback(() => {
-    if (!inline || window.innerWidth < 1180) {
-      setScalingFactor(
-        (window.innerWidth >= 1180 ? Math.min(window.innerWidth * 0.4, 720) : window.innerWidth) /
-          576,
-      );
-      setScaledHeight(Math.ceil(window.innerHeight / scalingFactor));
-      if (window.innerHeight != lastHeight && window.innerWidth == lastWidth) {
-        // This is probably the android keyboard opening
-        const delta = lastHeight - window.innerHeight;
-        if (delta > 0) widgetRef.current?.scrollTo(0, Math.ceil(delta / scalingFactor));
-        else widgetRef.current?.scrollTo(0, 0);
-      }
-      setLastWidth(window.innerWidth);
-      setLastHeight(window.innerHeight);
-    }
-  }, [
-    setScalingFactor,
-    setScaledHeight,
-    scalingFactor,
-    scaledHeight,
-    setLastWidth,
-    setLastHeight,
-    inline,
-  ]);
-
-  useEffect(() => scaleWidget, [widgetContext.open, scaleWidget]);
-
-  const debouncedScaleWidget = useDebouncedCallback(() => scaleWidget(), 1000, { maxWait: 1000 });
-
-  useEffect(() => {
-    if (!inline) {
-      window.addEventListener("resize", debouncedScaleWidget);
-
-      return () => {
-        window.removeEventListener("resize", debouncedScaleWidget);
-      };
-    }
-  }, [debouncedScaleWidget, inline]);
-
-  useEffect(() => {
-    scaleWidget();
-  }, [widgetContext, scaleWidget]);
-
-  return useMemo(() => ({ scaledHeight, scalingFactor }), [scaledHeight, scalingFactor]);
-};
-
 export const Widget = withStaticProps(
   async ({
     draftMode,
     inline,
     prefilled,
+    defaultPaymentType,
   }: {
     draftMode: boolean;
     inline?: boolean;
     prefilled?: PrefilledDistribution;
+    defaultPaymentType?: RecurringDonation;
   }) => {
     const result = await getClient(draftMode ? token : undefined).fetch<WidgetProps>(widgetQuery);
 
@@ -234,9 +110,10 @@ export const Widget = withStaticProps(
       },
       inline: inline ?? false,
       prefilled: prefilled ?? null,
+      defaultPaymentType: defaultPaymentType ?? RecurringDonation.NON_RECURRING,
     };
   },
-)(({ data, inline = false, prefilled }) => {
+)(({ data, inline = false, prefilled, defaultPaymentType }) => {
   const widget = data.result;
   const methods = data.result.methods;
 
@@ -278,6 +155,7 @@ export const Widget = withStaticProps(
   useQueryParamsPrefill({
     inline,
     causeAreas,
+    defaultPaymentType,
   });
 
   useDefaultPaymentMethodEffect(methods);
