+++ components/shared/components/Widget/components/hooks.ts
@@ -1,5 +1,5 @@
-import { useContext, useEffect, useRef } from "react";
-import { useDispatch } from "react-redux";
+import { useCallback, useContext, useEffect, useMemo, useRef, useState } from "react";
+import { useDispatch, useSelector } from "react-redux";
 import { useRouter } from "next/router";
 import {
   setCauseAreaPercentageShare,
@@ -13,6 +13,10 @@ import { WidgetContext } from "../../../../main/layout/layout";
 import { PrefilledDistribution } from "../../../../main/layout/WidgetPane/WidgetPane";
 import { CauseArea } from "../types/CauseArea";
 import { DistributionCauseArea } from "../types/DistributionCauseArea";
+import { WidgetProps } from "../types/WidgetProps";
+import { paymentMethodConfigurations } from "../config/methods";
+import { useDebouncedCallback } from "use-debounce";
+import { State } from "../store/state";
 
 interface UsePrefilledDistributionProps {
   inline: boolean;
@@ -100,9 +104,11 @@ export const usePrefilledSum = ({ inline }: { inline: boolean }) => {
 export const useQueryParamsPrefill = ({
   inline,
   causeAreas,
+  defaultPaymentType,
 }: {
   inline: boolean;
   causeAreas: CauseArea[] | undefined;
+  defaultPaymentType: RecurringDonation;
 }) => {
   const router = useRouter();
   const dispatch = useDispatch();
@@ -129,6 +135,10 @@ export const useQueryParamsPrefill = ({
       setWidgetContext({ ...widgetContext, open: true });
     }
   }, [inline, router.query, causeAreas, dispatch, setWidgetContext, widgetContext]);
+
+  useEffect(() => {
+    dispatch(setRecurring(defaultPaymentType));
+  }, [defaultPaymentType, dispatch]);
 };
 
 // Helper functions
@@ -183,4 +193,133 @@ const parseDistributionQueryParam = (distribution: string): PrefilledDistributio
       }),
     };
   });
+};
+
+/**
+ * Determine available recurring options based on the configured payment methods
+ */
+export const useAvailableRecurringOptions = (
+  paymentMethods: NonNullable<WidgetProps["methods"]>,
+) => {
+  const recurring = useMemo(
+    () =>
+      paymentMethods.some((method) => {
+        const configuration = paymentMethodConfigurations.find(
+          (config) => config.id === method._id,
+        );
+        return configuration?.recurringOptions.includes(RecurringDonation.RECURRING);
+      }),
+    [paymentMethods],
+  );
+
+  const single = useMemo(
+    () =>
+      paymentMethods.some((method) => {
+        const configuration = paymentMethodConfigurations.find(
+          (config) => config.id === method._id,
+        );
+        return configuration?.recurringOptions.includes(RecurringDonation.NON_RECURRING);
+      }),
+    [paymentMethods],
+  );
+
+  return useMemo(() => ({ recurring, single }), [recurring, single]);
+};
+
+/**
+ * Determine available payment methods based on the selected recurring option
+ */
+export const useAvailablePaymentMethods = (paymentMethods: NonNullable<WidgetProps["methods"]>) => {
+  const recurring = useSelector((state: State) => state.donation.recurring);
+
+  const availablePaymentMethods = useMemo(
+    () =>
+      paymentMethods.filter((method) => {
+        const configuration = paymentMethodConfigurations.find(
+          (config) => config.id === method._id,
+        );
+        return configuration?.recurringOptions.includes(recurring);
+      }),
+    [paymentMethods, recurring],
+  );
+
+  return availablePaymentMethods;
+};
+
+/**
+ * This effect is used to set the default payment method to single if recurring is not enabled
+ */
+export const useDefaultPaymentMethodEffect = (
+  paymentMethods: NonNullable<WidgetProps["methods"]>,
+) => {
+  const dispatch = useDispatch();
+  const recurring = useSelector((state: State) => state.donation.recurring);
+
+  const availableRecurringOptions = useAvailableRecurringOptions(paymentMethods);
+
+  useEffect(() => {
+    if (recurring === RecurringDonation.RECURRING && !availableRecurringOptions.recurring) {
+      dispatch(setRecurring(RecurringDonation.NON_RECURRING));
+    }
+  }, [recurring, availableRecurringOptions.recurring, dispatch]);
+};
+
+/**
+ * Scale the widget to fit the screen
+ */
+export const useWidgetScaleEffect = (
+  widgetRef: React.RefObject<HTMLDivElement>,
+  inline: boolean,
+) => {
+  const [widgetContext, setWidgetContext] = useContext(WidgetContext);
+  const [scalingFactor, setScalingFactor] = useState(1);
+  const [scaledHeight, setScaledHeight] = useState(979);
+  const [lastHeight, setLastHeight] = useState(979);
+  const [lastWidth, setLastWidth] = useState(400);
+
+  const scaleWidget = useCallback(() => {
+    if (!inline || window.innerWidth < 1180) {
+      setScalingFactor(
+        (window.innerWidth >= 1180 ? Math.min(window.innerWidth * 0.4, 720) : window.innerWidth) /
+          576,
+      );
+      setScaledHeight(Math.ceil(window.innerHeight / scalingFactor));
+      if (window.innerHeight != lastHeight && window.innerWidth == lastWidth) {
+        // This is probably the android keyboard opening
+        const delta = lastHeight - window.innerHeight;
+        if (delta > 0) widgetRef.current?.scrollTo(0, Math.ceil(delta / scalingFactor));
+        else widgetRef.current?.scrollTo(0, 0);
+      }
+      setLastWidth(window.innerWidth);
+      setLastHeight(window.innerHeight);
+    }
+  }, [
+    setScalingFactor,
+    setScaledHeight,
+    scalingFactor,
+    scaledHeight,
+    setLastWidth,
+    setLastHeight,
+    inline,
+  ]);
+
+  useEffect(() => scaleWidget, [widgetContext.open, scaleWidget]);
+
+  const debouncedScaleWidget = useDebouncedCallback(() => scaleWidget(), 1000, { maxWait: 1000 });
+
+  useEffect(() => {
+    if (!inline) {
+      window.addEventListener("resize", debouncedScaleWidget);
+
+      return () => {
+        window.removeEventListener("resize", debouncedScaleWidget);
+      };
+    }
+  }, [debouncedScaleWidget, inline]);
+
+  useEffect(() => {
+    scaleWidget();
+  }, [widgetContext, scaleWidget]);
+
+  return useMemo(() => ({ scaledHeight, scalingFactor }), [scaledHeight, scalingFactor]);
 };
